Partial implementation of getRandomState/setRandomState
Benefitted from Koray's help to modify the headers and their order of
inclusions, so as to allow Random to access Tensor's functions. Thanks
a lot Koray!
