+++ lib/TH/TH.h
@@ -10,7 +10,6 @@
 
 #include "THVector.h"
 #include "THLogAdd.h"
-#include "THRandom.h"
 #include "THStorage.h"
 #include "THTensor.h"
 #include "THTensorApply.h"
