+++ lib/TH/THTensor.c
@@ -2,16 +2,12 @@
 #include "THVector.h"
 #include "THBlas.h"
 #include "THLapack.h"
-#include "THRandom.h"
 #include "THTensorDimApply.h"
 
 #include "generic/THTensor.c"
 #include "THGenerateAllTypes.h"
 
 #include "generic/THTensorCopy.c"
-#include "THGenerateAllTypes.h"
-
-#include "generic/THTensorRandom.c"
 #include "THGenerateAllTypes.h"
 
 #include "generic/THTensorMath.c"
